requirements-app.txt
fastapi==0.115.12
markdown-it-py==3.0.0
uvicorn==0.34.2
pywebview==5.4
pyinstaller==
